requirements.txt
aiohttp
enkacard
pymongo
python-decouple
whatsapp_api_client_python

#Snyk pins 
certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
dnspython>=2.6.0rc1 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
